movign ErrorAndWarnings to error_helpers
