+++ pkg/error_helpers/errors.go
@@ -1,6 +1,7 @@
 package error_helpers
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/turbot/steampipe/pkg/constants"
@@ -8,3 +9,4 @@ import (
 
 var MissingCloudTokenError = fmt.Errorf("Not authenticated for Steampipe Cloud.\nPlease run %s or setup a token.", constants.Bold("steampipe login"))
 var InvalidCloudTokenError = fmt.Errorf("Invalid token.\nPlease run %s or setup a token.", constants.Bold("steampipe login"))
+var InvalidStateError = errors.New("invalid state")
